ReadCompressed(ElementID) -> Read(ElementID) WriteCompressed(ElementID) -> Write(ElementID) For all instances.
We are going to do the "sacrifice" of not compressing element ids in rare packets: lua rpcs, element data, triggers, world rpcs, etc... in change of saving bandwidth in puresync packets.

It actually is not a big sacrifice because as element ids grow non stop, they lose the capability of being synced compressed after getting over 65535 and actually waste an extra bit.