Fixes for vehicles falling through the map.
Fix #1 - Stop vehicles falling through the San Andreas map - anything that falls under the map and will be warped back to the original spawn position for 2 retrys at which point it is considered unfixable and will be warped to the nearest path node.

Fix #2 - Stop vehicles falling through custom objects - Added a wait for the ground to load to Create.